Expenses Animal table

+++ app/src/main/java/com/zaroslikov/fermacompose2/data/animal/AnimalTable.kt
@@ -28,7 +28,7 @@ data class AnimalTable(
     val image:String,
     val arhiv: Boolean,
     val price: Double,
-    val foodDay:Int,
+    val foodDay:Double,
     @ColumnInfo(name = "idPT")
     val idPT: Int,
     )
